add a follow/unfollow button next to the userlink
So this adds a `follow` or `unfollow` link that triggers the pertinent
event when clicked. So the approach I've taken is to add a method to
every userlink, but I don't know if it is the correct approach, the way
it works now is it'll check if it is you, if it isn't it'll check if
you're following or not and then put the link with the action needed
right beside it.

What's cool is that since we're using mercury every other link auto
updates if you move from following to unfollowing and viceversa. So
that's neat.