add bubble notification to give the user feedback
Exposes a model `bubble` with the properties msg, type and show to
display notifications much like the disconnected from the back end one
usually you don't want to manage the `show` everywhere, instead you want
to have it in one place only and be automatic.

I am using it in `follow` and `unfollow` events only but in theory it
could be used for any event that needs notifying. There is one caveat
only: each notification overwrites the previous one, so there is no
stacking notifications.